backend_server/requirements.txt
Django==1.11.7
django-extensions==1.9.7
ipython==6.2.1
requests==2.18.4
mysqlclient==1.3.12
PyJWT==1.5.3
